collapse models/ organization to two deep
