--- src/opinf/models/monolithic/__init__.py
@@ -1,7 +1,0 @@
-# models/monolithic__init__.py
-"""Monolithic dynamical systems models."""
-
-from .nonparametric import *
-from .parametric import *
-
-__all__ = nonparametric.__all__ + parametric.__all__
