--- src/opinf/models/monolithic/nonparametric/__init__.py
@@ -1,8 +1,0 @@
-# models/monolithic/nonparametric/__init__.py
-"""Monolithic nonparametric dynamical systems models."""
-
-from . import _base, _frozen
-from ._public import *
-
-
-__all__ = _public.__all__
