--- src/opinf/models/monolithic/parametric/__init__.py
@@ -1,8 +1,0 @@
-# models/monolithic/nparametric/__init__.py
-"""Monolithic parametric dynamical systems models."""
-
-from . import _base
-from ._public import *
-
-
-__all__ = _public.__all__
