+++ src/opinf/models/multi/__init__.py
@@ -1,0 +1,7 @@
+# models/multi/__init__.py
+"""Multilithic dynamical systems models."""
+
+from ._nonparametric import *
+from ._parametric import *
+
+__all__ = _nonparametric.__all__ + _parametric.__all__
