+++ src/opinf/models/multi/_base.py
@@ -1,0 +1,4 @@
+# models/multi/_base.py
+"""Abstract base class for multilithic dynamical systems models."""
+
+__all__ = []
