+++ src/opinf/models/multi/_nonparametric.py
@@ -1,0 +1,4 @@
+# models/multi/_nonparametric.py
+"""Nonparametric multilithic dynamical systems models."""
+
+__all__ = []
