+++ src/opinf/models/multi/_parametric.py
@@ -1,0 +1,4 @@
+# models/multi/_parametric.py
+"""Parametric multilithic dynamical systems models."""
+
+__all__ = []
