--- src/opinf/models/multilithic/__init__.py
@@ -1,8 +1,0 @@
-# models/multilithic__init__.py
-"""Multilithic dynamical systems models."""
-
-from .nonparametric import *
-from .parametric import *
-
-
-__all__ = nonparametric.__all__ + parametric.__all__
