--- src/opinf/models/multilithic/nonparametric/__init__.py
@@ -1,8 +1,0 @@
-# models/multilithic/nonparametric/__init__.py
-"""Multilithic nonparametric dynamical systems models."""
-
-from . import _base, _frozen
-from ._public import *
-
-
-__all__ = _public.__all__
