--- src/opinf/models/multilithic/nonparametric/_base.py
@@ -1,5 +1,0 @@
-"""Base class for multilithic nonparametric models."""
-
-__all__ = []
-
-# from .._base import _MultilithicModel
