--- src/opinf/models/multilithic/nonparametric/_frozen.py
@@ -1,31 +1,0 @@
-# models/multilithic/nonparametric/_frozen.py
-"""Evaluations of multilithic parametric model classes with fit() disabled."""
-
-__all__ = []
-
-# from ._public import (
-#     SteadyModelMulti, DiscreteModelMulti, ContinuousModelMulti)
-
-
-# class _FrozenMixin:
-#     """Mixin for evaluations of parametric models (disables fit())."""
-
-#     pass  # P
-
-
-# class _FrozenSteadyModelMulti(_FrozenMixin, SteadyModelMulti):
-#     """Steady-state model that is the evaluation of a parametric model."""
-
-#     pass  # pragma: no cover
-
-
-# class _FrozenDiscreteModelMulti(_FrozenMixin, DiscreteModelMulti):
-#     """Discrete-time model that is the evaluation of a parametric model."""
-
-#     pass  # pragma: no cover
-
-
-# class _FrozenContinuousModelMulti(_FrozenMixin, ContinuousModelMulti):
-#     """Continuous-time model that is the evaluation of a parametric model."""
-
-#     pass  # pragma: no cover
