--- src/opinf/models/multilithic/nonparametric/_public.py
@@ -1,10 +1,0 @@
-"""Public multilithic nonparametric models."""
-
-__all__ = []
-#     "SteadyModelMulti",
-#     "DiscreteModelMulti",
-#     "ContinuousModelMulti",
-# ]
-
-
-# from ._base import _NonparametricMultilithicModel
