--- src/opinf/models/multilithic/parametric/__init__.py
@@ -1,8 +1,0 @@
-# models/multilithic/nonparametric/__init__.py
-"""Multilithic parametric dynamical systems models."""
-
-from . import _base
-from ._public import *
-
-
-__all__ = _public.__all__
