--- src/opinf/models/multilithic/parametric/_base.py
@@ -1,7 +1,0 @@
-# models/multilithic/parametric/_base.py
-"""Base class for multilithic parametric dynamical systems models."""
-
-__all__ = []
-
-
-# class _ParametricMultilithicModel
