--- src/opinf/models/multilithic/parametric/_public.py
@@ -1,6 +1,0 @@
-# models/multilithic/parametric/_public.py
-"""Public multilithic parametric model classes."""
-
-__all__ = []
-
-# from ._base import _ParametricMultilithicModel
