+++ tests/models/mono/test_parametric.py
@@ -1,0 +1,6 @@
+# models/mono/test_parametric.py
+"""Tests for models.mono._parametric."""
+
+import opinf
+
+_module = opinf.models.mono._parametric
